pyproject.toml
[tool.poetry]
name = "inkmanager"
version = "0.1.0"
description = ""
authors = ["Lucas Ferreira <lucasferreiraek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.14.0"
ruff = "^0.7.1"
testcontainers = "^4.8.2"
factory-boy = "^3.3.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev inkmanager/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=inkmanager -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
